+++ jvm-packages/xgboost4j-spark/src/test/scala/ml/dmlc/xgboost4j/scala/spark/XGBoostCommunicatorRegressionSuite.scala
@@ -47,11 +47,6 @@ class XGBoostCommunicatorRegressionSuite extends AnyFunSuite with PerTest {
     val model2 = new XGBoostClassifier(xgbSettings ++ Map("rabit_ring_reduce_threshold" -> 1))
       .fit(training)
 
-    // assert(Communicator.communicatorEnvs.asScala.size > 3)
-    // Communicator.communicatorEnvs.asScala.foreach( item => {
-    //   if (item._1.toString == "rabit_reduce_ring_mincount") assert(item._2 == "1")
-    // })
-
     val prediction2 = model2.transform(testDF).select("prediction").collect()
     // check parity w/o rabit cache
     prediction1.zip(prediction2).foreach { case (Row(p1: Double), Row(p2: Double)) =>
@@ -70,10 +65,6 @@ class XGBoostCommunicatorRegressionSuite extends AnyFunSuite with PerTest {
 
     val model2 = new XGBoostRegressor(xgbSettings ++ Map("rabit_ring_reduce_threshold" -> 1)
     ).fit(training)
-    // assert(Communicator.communicatorEnvs.asScala.size > 3)
-    // Communicator.communicatorEnvs.asScala.foreach( item => {
-    //   if (item._1.toString == "rabit_reduce_ring_mincount") assert(item._2 == "1")
-    // })
     // check the equality of single instance prediction
     val prediction2 = model2.transform(testDF).select("prediction").collect()
     // check parity w/o rabit cache
