--- tests/ci_build/build_mock_cmake.sh
@@ -1,10 +1,0 @@
-#!/usr/bin/env bash
-set -e
-
-rm -rf build
-mkdir build
-cd build
-cmake -DRABIT_MOCK=ON -DCMAKE_VERBOSE_MAKEFILE=ON ..
-make clean
-make -j$(nproc)
-cd ..
