+++ tests/cpp/common/test_io.cc
@@ -15,8 +15,8 @@ namespace xgboost::common {
 TEST(MemoryFixSizeBuffer, Seek) {
   size_t constexpr kSize { 64 };
   std::vector<int32_t> memory( kSize );
-  rabit::utils::MemoryFixSizeBuffer buf(memory.data(), memory.size());
-  buf.Seek(rabit::utils::MemoryFixSizeBuffer::kSeekEnd);
+  MemoryFixSizeBuffer buf(memory.data(), memory.size());
+  buf.Seek(MemoryFixSizeBuffer::kSeekEnd);
   size_t end = buf.Tell();
   ASSERT_EQ(end, kSize);
 }
